Create README.md
Due to the fact that Chinese mothers do not trust lokale made infant formula milk the marked for this product made in Europe,Scandinavia is huge.We are at the moment developing an infant formula in Norway with Shanghai as our target( 190 mill people) Through the Norwegian/ Chinese female entrepreneur Chen Suofen´s network in China the Arctic babfood infant formula be distributed via social networks and online shopping. The product is due to be shipped to China in early February 2014. We need money to get the first order going.